window: Port the base of the window to Glade
This also puts the header bar into a HdyStitleBar and a GtkStack as
these will be needed when the header bar will be split into one per
page.

This also reorders the pages so we could use an order-dependant stack
animation and avoid having to change it depending on the pages.