conversation: Reverse messages in each log
While b7360ff112a84437c3b68072c5bbe1388fc00c67 took care of not
discarding new messges the reversion of the list resulted in reversed
entries for each log file (so the overall history was in the correct
order but each converstion was reversed).

We go this way around instead of using g_list_append() since the later
is slower.

Fix that by going through each log back to front.