BuddyList: Clean up code and remove manual refresh for buddylist
Create the Chats and Contacts Listboxes before they are needed. Also we
can connect to all events imiditly and we shouldn't wait for the show
event to do so. This moves most of the code from the show ops to init.

Libpurple notifies about buddy changes, there shouldn't be a need
for a manuel refresh.

This also removes code which isn't used anymore.