Simplify managing accounts created
The ‘ui_data’ variable in PurpleAccount is unused.  Let’s reuse
it to keep track of the GObject the PurpleAccount belongs to.

This helps us keep track of those objects and get them when
required without iterating the account list (which is O(n),
though not much concern as we don’t expect anyone to have
hundreds of accounts).