manager: enable notifications on first message
When receiving a message from anyone for the first time, notifications
aren't enabled by default, meaning messages can be missed if the user
doesn't check chatty on a regular basis.

This patch enables notifications each time a chat is started, either
initiated by the user or created because a message has been received.