Remove chatty-folks
chatty_contact_get_individual() always returned ‘NULL’ and thus
‘chatty_folks_set_purple_buddy_data()’ code was never run.

Also, if a buddy has an evolution contact, the avatar from the
contact is used if available.  So we don't have to get the avatar
from folks and save it to the buddy.

And thus we don’t need chatty-folks as that was the sole purpose
of the file.