window: ensure received SMS is displayed
GtkFilterListModel doesn't have a signal named "items-changed", we
therefore need to connect to the GListModel instead of the
GtkFilterListModel.

Furthermore, when an SMS is received from an addressbook contact, the
"chatty-autojoin" property is not set, causing this SMS to be hidden
unless using the search function. In order to prevent that, the check on
"chatty-autojoin" is no more performed on SMS messages

Closes #305