history: Fix possible deadlock
history is a singleton object.  It's created and finalized only once in
an application's lifetime.  But in tests, the same object may be reused
after closing database, which can result in races that lock the GTask to
never complete.

Free object data in dispose and explicitly run dispose when the database
is closed so that we don't have wait for GObject to free data which may
be delayed.