manager: Remove redundant code
"conversation-created" and "chat-joined" signals are emitted
immediately after the conversation is created.

When the conversation is created, it is handled by chatty_conv_new(),
which adds the PurpleConversation to the chat list.  So what ever
"conversation-created" and "chat-joined" callbacks where doing are
already done by chatty_conv_new().  So let's remove these signals
and their handlers as nothing is done there.