chat-view: Don't let Return key insert newline if set to send message
If Return key press is set to send the inserted message, never insert
a newline when the key is pressed.

This was happening when the message input is empty and thus the Return
key press wasn't claimed to send the message.

If Return key is pressed with the message input empty, do an arror bell
instead

Fixes #391