chat: Fix setting ui_data when buddy/chat is already set
If ChattyChat has a PurpleBuddy or PurpleChat already set,
the `ui_data` of the PurpleConversation wasn't updated
when a conv is set for the chat.

So instead, check for PurpleBuddy/PurpleChat after the `ui_data'
is updated.  Also, remove previously set `ui_data` for the conv
already set in the chat, if any.