manager: Remove some lists used to store chats
We now make use of ui_data for tracking various objects.  And
so we no longer have to lookup in lists to find the objects.

Don't differentiate between IM and chats.  Use a single liststore
for any type of chats, and so we no longer require im_list, and
related helper models.