history: Implement new Database design
This commit implements new design for the database.

Previously we had two tables, chatty_chat and chatty_im, with
the only difference b/w them is the 'room' column.  The
'account', 'who', and 'room'  values were repeated in several
rows.

This commit normalizes them by adding new tables for handling
'account', 'room' and 'who' and moving the common content to
'messages' table.  This commit also adds additional tables to
store files and other media.

The data from old database is migrated to the new database with
all possible information guessed from the available data (like,
we were not saving the protocol of a chat, best effort is made to
guess and fill in these information on migration)

The old database is backup before migration so that the data
won't be lost in case of any fatal errors.

Tests are also updated accordingly.

Required for #407 #215 #30

Fixes #382