utils: Improve parsing phone numbers
If a number is possible to be parsed with a country prefix (eg: +1)
always do so.  This will help the numbers to be the same regardless
of what the user inputs, and thus help the messages to be in the
same chat thread.

This commit also fixes international phone numbers with country
prefix being converted to local numbers when not prefixed with '+'.

As a side effect, possible phone numbers are converted to E.164.

Partly fixes #382