manager: Don't load second last message twice
When the conversation is created, the corresponding chat may
already have fetched the last message.  So load a message
only if no message is loaded.

Otherwise, if a user opens a chat, the second last message
is loaded when the conversation is created.  Also, when the
chat is loaded the recent history is loaded, and since both
these requests are done async, the second last message is
loaded twice.