build: Test ui files
Minimum version has been bumped to 0.46 as it's required
for both_libraries()

both_libraries() is used because we need shared library for
LD_PRELOAD and static one to create the binary.

Since chatty-avatar is used in .ui file, it has been moved to
libsrc

library name has been renamed from 'libchatty' to 'chatty' as
meson prefixes a 'lib' to the name which resulted in the
library name to be 'liblibchatty'.

Also, run tests with xvfb-run as gtk-builder-tool requires a display