matrix-api: Increase connections limit
The default per host connection limit is 2. Since `/sync` API is a blocking
one, we have just 1 another connection for use, for which the request can
be delayed if there is already a pending request.

So increase the limit to 6 (which is a lower common value used by many browsers)