Implement new notifications
This is a generic notifications that doesn't use any Purple
APIs, so that this can be used with other providers (like matrix).

Only the features used in chatty-notify.c is implemented here.
chatty-notify.c was used only to show message notifications,
and so the rest are not included.

The notifications are also shown after some delay, and multiple
message notifications are squashed into one when they are
received within a short timespan (currently 300 milliseconds)

Fixes #389