history: Return task error on failures
In case of some errors, we were simply returning
(sometimes with warning) without returning the
associated tasks, resulting in the functions being
blocked on some errors when used as sync.

So Let's handle much of the errors so that
associated tasks returns on error.

Fixes https://source.puri.sm/Librem5/chatty/-/issues/456#note_141787