message-row: Fix being stuck in the loop updating quote
'>' can happen midst a sentence where the line is not considered
a quote.  In this case, the 'end' was not updated resulting a
search for '>' happening again and again in the same string.

So in that cases, move forward one letter (actually one byte)
so as to begin next search from the following character.

Reported by @evangelos.tzaras